requirements.txt
Flask
gunicorn
numpy 
pandas
sklearn
pyYAML
-e .